[libromdata] WiiWAD: Use unsigned int for the title version, even though it's uint16_t.
C integer promotion rules result in the uint16_t becoming int when
performing operations on it, which causes a -Wformat warning.

Found in the AppVeyor MinGW-w64 build, but not in the local MinGW-w64
build...

+++ src/libromdata/Console/WiiWAD.cpp
@@ -971,7 +971,7 @@ int WiiWAD::loadFieldData(void)
 	}
 
 	// Title version.
-	const uint16_t title_version = be16_to_cpu(tmdHeader->title_version);
+	const unsigned int title_version = be16_to_cpu(tmdHeader->title_version);
 	d->fields->addField_string(C_("WiiWAD", "Title Version"),
 		rp_sprintf("%u.%u (v%u)", title_version >> 8, title_version & 0xFF, title_version));
